:truck: rename: move dto package
